test/shim_test: relocate hw_type and dev_filter entries to new dev_filter.h
