+++ test/shim_test/shim_test.cpp
@@ -8,6 +8,7 @@
 
 #include "multi_threads.h"
 #include "dev_info.h"
+#include "dev_filter.h"
 #include "io_param.h"
 #include "hwctx.h"
 #include "speed.h"
@@ -134,21 +135,6 @@ usage(const std::string& prog)
   std::cout << std::endl;
 }
 
-// Hardware platform and driver type enums for test case filtering.
-// Any device can belong to only 1 hw_type, no overlap.
-enum hw_type {
-  non_npu,  // non-xdna device
-  npu1,     // AIE2, device_id 0x1502
-  npu4,     // AIE2, device_id 0x17f0
-  npu3,     // AIE4 classic, device_id 0x17f1 / 0x1b0a (classic only)
-  npu3vf,   // AIE4 VF, device_id 0x17f3 / 0x1b0c (SRIOV VF only)
-  ve2,      // VE2 edge, device_id 0xb052
-};
-
-enum drv_type {
-  amdxdna,  // native amdxdna kernel driver
-};
-
 // Definition of one test case
 struct test_case {
   const std::string name;
@@ -169,127 +155,11 @@ std::vector<int> test_passed;
 std::vector<int> test_skipped;
 std::vector<int> test_failed;
 
-// Per-enum hw filter entry: nullptr check for non_npu (special case)
-struct hw_filter_entry {
-  hw_type hw;
-  bool (*check)(device::id_type, device*);
-};
-
-struct drv_filter_entry {
-  drv_type drv;
-  bool (*check)(device::id_type, device*);
-};
-
-// Per-enum hw check functions — implement device ID logic directly
-bool
-dev_filter_is_npu1(device::id_type id, device* dev)
-{
-  return device_query<query::pcie_device>(dev) == npu1_device_id;
-}
-
-bool
-dev_filter_is_npu4(device::id_type id, device* dev)
-{
-  return device_query<query::pcie_device>(dev) == npu4_device_id;
-}
-
-bool
-dev_filter_is_npu3(device::id_type id, device* dev)
-{
-  auto device_id = device_query<query::pcie_device>(dev);
-  return device_id == npu3_device_id || device_id == npu3a_device_id;
-}
-
-bool
-dev_filter_is_npu3vf(device::id_type id, device* dev)
-{
-  auto device_id = device_query<query::pcie_device>(dev);
-  return device_id == npu3_device_id1 || device_id == npu3a_device_id1;
-}
-
-bool
-dev_filter_is_ve2(device::id_type id, device* dev)
-{
-  return device_query<query::pcie_device>(dev) == npu_ve2_device_id;
-}
-
-// hw_filter_table: non_npu is entry 0 with nullptr check (special case handled by
-// dev_filter_is_non_npu). All other entries have a dedicated check function.
-// When adding a new hw_type, add it here and dev_filter_is_non_npu auto-updates.
-static const hw_filter_entry hw_filter_table[] = {
-  { non_npu, nullptr              },
-  { npu1,    dev_filter_is_npu1   },
-  { npu4,    dev_filter_is_npu4   },
-  { npu3,    dev_filter_is_npu3   },
-  { npu3vf,  dev_filter_is_npu3vf},
-  { ve2,     dev_filter_is_ve2    },
-};
-
-// driver checker
-bool
-drv_filter_is_amdxdna(device::id_type id, device* dev)
-{
-  query::sub_device_path::args query_arg = {std::string(""), 0};
-  auto sysfs = device_query<query::sub_device_path>(dev, query_arg);
-  auto drv_path = std::filesystem::read_symlink(sysfs + "/driver");
-  return drv_path.filename() == "amdxdna";
-}
-
-static const drv_filter_entry drv_filter_table[] = {
-  { amdxdna, drv_filter_is_amdxdna },
-};
-
-// dev_filter_is_non_npu: iterates hw_filter_table from entry 1 onward,
-// returns true if none of the known NPU checkers match.
-bool
-dev_filter_is_non_npu(device::id_type id, device* dev)
-{
-  for (size_t i = 1; i < std::size(hw_filter_table); i++) {
-    if (hw_filter_table[i].check(id, dev))
-      return false;
-  }
-  return true;
-}
-
-// Overall evaluation functions
-bool
-match_hw(const std::vector<hw_type>& hw_list, device::id_type id, device* dev)
-{
-  if (hw_list.empty())
-    throw std::runtime_error("match_hw: hw_list must not be empty");
-  for (auto hw : hw_list) {
-    if (hw == non_npu) {
-      if (dev_filter_is_non_npu(id, dev))
-        return true;
-      continue;
-    }
-    for (const auto& entry : hw_filter_table) {
-      if (entry.hw == hw && entry.check(id, dev))
-        return true;
-    }
-  }
-  return false;
-}
-
-bool
-match_drv(const std::vector<drv_type>& drv_list, device::id_type id, device* dev)
-{
-  if (drv_list.empty())
-    return true;
-  for (auto drv : drv_list) {
-    for (const auto& entry : drv_filter_table) {
-      if (entry.drv == drv && entry.check(id, dev))
-        return true;
-    }
-  }
-  return false;
-}
-
 static void TEST_async_error_io_any(device::id_type id, std::shared_ptr<device>& sdev, arg_type& arg)
 {
   if (dev_filter_is_npu4(id, sdev.get()))
     TEST_async_error_io(id, sdev, arg);
-  else if (dev_filter_is_npu3(id, sdev.get()) || dev_filter_is_npu3vf(id, sdev.get()))
+  else if (dev_filter_is_aie4(id, sdev.get()))
     TEST_async_error_aie4_io(id, sdev, arg);
   else
     throw std::runtime_error("async error io test: device is neither NPU4 nor NPU3");
@@ -794,7 +664,7 @@ TEST_create_destroy_max_context(device::id_type id, std::shared_ptr<device>& sde
   // XDNA driver by default supports maximum 6 contexts on npu1, 128 on npu3, and 16 on npu4
   if (dev_filter_is_npu1(id, dev))
     num_ctx = 6;
-  else if (dev_filter_is_npu3(id, dev) || dev_filter_is_npu3vf(id, dev))
+  else if (dev_filter_is_aie4(id, dev))
     num_ctx = 128;
   else
     num_ctx = 16;
@@ -819,7 +689,7 @@ TEST_multi_context_io_test(device::id_type id, std::shared_ptr<device>& sdev, ar
   const std::array<int, 3> ctx = [&]() {
     if (dev_filter_is_npu1(id, dev))
       return std::array<int, 3>{2, 4, 6};
-    if (dev_filter_is_npu3(id, dev) || dev_filter_is_npu3vf(id, dev))
+    if (dev_filter_is_aie4(id, dev))
       return std::array<int, 3>{4, 8, 32};
     return std::array<int, 3>{4, 8, 16};
   }();
@@ -835,7 +705,7 @@ TEST_create_free_debug_bo(device::id_type id, std::shared_ptr<device>& sdev, arg
 {
   auto dev = sdev.get();
   auto boflags = XRT_BO_FLAGS_CACHEABLE;
-  auto ext_boflags = (dev_filter_is_npu3(id, dev) || dev_filter_is_npu3vf(id, dev)) ?
+  auto ext_boflags = dev_filter_is_aie4(id, dev) ?
     (XRT_BO_USE_UC_DEBUG << 4) : (XRT_BO_USE_DEBUG << 4);
   auto size = static_cast<size_t>(arg[0]);
 
@@ -1503,7 +1373,7 @@ struct telemetry_probe {
 void
 TEST_query_telemetry(device::id_type id, std::shared_ptr<device>& sdev, arg_type& arg)
 {
-  uint32_t type = (dev_filter_is_npu3(id, sdev.get()) || dev_filter_is_npu3vf(id, sdev.get())) ?
+  uint32_t type = dev_filter_is_aie4(id, sdev.get()) ?
     AIE4_TELEMETRY_PERF_COUNTER : 0;
 
   auto probe = fork_query<telemetry_probe>(sdev.get(), [type](int fd) {
@@ -1587,7 +1457,7 @@ TEST_query_frame_boundary_preempt(device::id_type id, std::shared_ptr<device>& s
       "ioctl(GET_FRAME_BOUNDARY_PREEMPT_STATE) failed: " + std::string(std::strerror(probe.err)));
   if (probe.state > 1)
     throw std::runtime_error("GET_FRAME_BOUNDARY_PREEMPT_STATE returned non-boolean state");
-  if ((dev_filter_is_npu3(id, sdev.get()) || dev_filter_is_npu3vf(id, sdev.get())) &&
+  if (dev_filter_is_aie4(id, sdev.get()) &&
       probe.state != 1)
     throw std::runtime_error("aie4 GET_FRAME_BOUNDARY_PREEMPT_STATE expected enabled");
 }
